fix test script
